Revert "Checks of input"
