+++ group/assignment6/assignment6.py
@@ -9,8 +9,7 @@ Car #1 is in the lot #0, car #2 is in the lot #1, etc.
 """
 
 from parking_class import Parking
-from path_finder import PathFinder
-from helpers import apply_moves, compute_move, input_check
+
 
 def compute_moves(start_state, end_state):
     """
@@ -19,6 +18,7 @@ def compute_moves(start_state, end_state):
     For every lot number we look up which car should be placed in this lot.
     The current car is moved to the empty lot, the correct car is moved
     to the current lot.
+
     :param start_state: order of cars in the start of the rearrangement
     :param end_state: order of cars after rearrangement
     :yields: move steps. Each move is represented as a tuple with two indeces,
@@ -27,7 +27,7 @@ def compute_moves(start_state, end_state):
     """
     if len(start_state) != len(end_state):
         raise IndexError('The start state and end state have different lengths.')
-    input_check(start_state, end_state)
+
     parking = Parking(start_state.copy())
     for lot in range(len(parking)):
         end_car = end_state[lot]
@@ -40,7 +40,7 @@ def compute_moves(start_state, end_state):
 
 def compute_efficient_moves(start_state, end_state):
     """
-    More efficiently computes the shortest sequence of moves that are required to rearrange 
+    More efficiently computes a sequence of moves that are required to rearrange 
     cars from the given start state to the end state. We look up which lot is 
     empty and which car should be in this lot. Then we move the required car
     to the empty lot and repeat the procedure. 
@@ -48,6 +48,7 @@ def compute_efficient_moves(start_state, end_state):
     with the misplaced car and move the car from that lot to the empty one. 
     Then we perform the initial steps. When all the cars are on their places
     (misplaced_car == len(current_state)), algorithm is finished.
+
     :param start_state: order of cars in the start of the rearrangement
     :param end_state: order of cars after rearrangement
     :yields: move steps. Each move is represented as a tuple with two indeces,
@@ -56,8 +57,6 @@ def compute_efficient_moves(start_state, end_state):
     """
     if len(start_state) != len(end_state):
         raise IndexError('The start state and end state have different lengths.')
-    input_check(start_state, end_state)
-    
     parking = Parking(start_state.copy())
     misplaced_car_lot = 0
     
@@ -76,65 +75,27 @@ def compute_efficient_moves(start_state, end_state):
             yield parking.move_to_empty_lot(parking.get_car(misplaced_car_lot))
 
 
-def compute_moves_with_constraints(start_state, end_state, constraints):
+def apply_moves(start_state, moves):
     """
-    Computes moves sequence under given constraints.
-    :param start_state: order of cars in the start of the rearrangement
-    :param end_state: order of cars after rearrangement
-    :param constraints: map from the parking lot to a tuple of the allowed cars
-    :yields: move steps. Each move is represented as a tuple with two indeces,
-             the 1st index is the number of lot from which we move the car,
-             the 2nd index is the number of lot to which we move the car
-    """
-    input_check(start_state, end_state)
-    path_finder = PathFinder(tuple(start_state), tuple(end_state), constraints)
-    paths = path_finder.find_all_paths()
-    path = path_finder.decode_path(next(paths))
-    for i in range(1, len(path)):
-        yield compute_move(path[i - 1], path[i])
+    Applies the sequence of moves to a given state.
 
-
-def compute_all_moves(start_state, end_state):
-    """
-    Computes alll possible ways to rearrange the cars from start_state to end_state.
     :param start_state: order of cars in the start of the rearrangement
-    :param end_state: order of cars after rearrangement
-    :yields: a list of move steps. Each move is represented as a tuple with two indeces,
-             the 1st index is the number of lot from which we move the car,
-             the 2nd index is the number of lot to which we move the car
+    :param moves: generator, sequence of moves that we need to apply.
+                  Each move is represented as a tuple with two indices,
+                  the 1st index is the number of lot from which we move the car,
+                  the 2nd index is the number of lot to which we move the car
+    :returns: end state, list of car numbers.
     """
-    input_check(start_state, end_state)
-    path_finder = PathFinder(tuple(start_state), tuple(end_state), constraints={}) 
-    paths = path_finder.find_all_paths()
-    for path in paths:
-        decoded_path = path_finder.decode_path(path)
-        moves_sequence = list()
-        for i in range(1, len(decoded_path)):
-            moves_sequence.append(compute_move(decoded_path[i - 1], decoded_path[i]))
-        yield moves_sequence
-    
+    parking = Parking(start_state.copy())
+    for move in moves:
+        move_from, move_to = move
+        parking.move_to_empty_lot(parking.get_car(move_from))
+    return parking.get_state()
+
 
 if __name__ == '__main__':
-    start_state = [1, 2, 3] 
-    end_state = [3, 1, 2]
-
-    print("Computing a moves sequence for the main problem...")
+    start_state = [1, 2, 0, 3] 
+    end_state = [3, 1, 2, 0]
     moves = list(compute_moves(start_state, end_state))
     print(moves)
-    
-    print("Computing the shortest moves sequence...")
-    moves = list(compute_efficient_moves(start_state, end_state))
-    print(moves)
-
-    print("Computing a moves sequence under given constraints...")
-    constraints = {0: (1, 3), 1: (0, 1, 2, 3), 2: (0, 1, 2, 3), 3: (0, 1, 2, 3)}
-    moves = list(compute_moves_with_constraints(start_state, end_state, constraints))
-    print(moves)
-
-    print("Computing all moves sequences between two parking states...")
-    moves_sequences = list(compute_all_moves(start_state, end_state))
-    n_sequences = len(moves_sequences)
-    n = min(5, n_sequences)
-    print("Found {} possible moves sequences, printing first {} moves sequences:".format(n_sequences, n)) 
-    for moves in moves_sequences[:n]:
-        print(moves)
+    print(apply_moves(start_state, moves))
